Add function of bigPicture generating to popup.js
