+++ js/main.js
@@ -1,5 +1,5 @@
 import {postItems} from './data-posts-generator.js';
-const dataPosts = postItems();
 import {addPictures} from './picture-generator.js';
+//import {openPictureModal, closePictureModal} from './popup.js';
+const dataPosts = postItems();
 addPictures(dataPosts);
-import {openPictureModal, closePictureModal} from './popup.js';
