+++ js/picture-generator.js
@@ -15,37 +15,7 @@ const addPictures = (pictures) => {
     const allPictures = createPicture(picture);
     allPictures.addEventListener('click', (event) => {
       event.preventDefault();
-      console.log(picture);
       openPictureModal(picture);
-      const bigPictureSrc = document.querySelector('.big-picture__img img');
-      const bigPictureLikes = document.querySelector('.likes-count');
-      const bigPictureCommentsCount = document.querySelector('.comments-count');
-      const bigPictureDescription = document.querySelector('.social__caption');
-      const bigPictureCommentsBlock = document.querySelector('.social__comments');
-      bigPictureCommentsBlock.innerHTML = '';
-      picture.comments.forEach((element) => {
-        const commentElement = document.createElement('li');
-        const commentElementImage = document.createElement('img');
-        const commentElementDescription = document.createElement('p');
-
-        commentElement.className = 'social__comment';
-        commentElementImage.className = 'social__picture';
-        commentElementImage.src = element.avatar;
-        commentElementImage.alt = element.name;
-        commentElementImage.width = 25;
-        commentElementImage.heigth = 25;
-        commentElementDescription.className = 'social__text"';
-        commentElementDescription.innerHTML = element.message;
-        commentElement.appendChild(commentElementImage);
-        commentElement.appendChild(commentElementDescription);
-        bigPictureCommentsBlock.appendChild(commentElement);
-      });
-
-      bigPictureSrc.src = picture.url;
-      bigPictureLikes.textContent = picture.likes;
-      bigPictureCommentsCount.textContent = picture.comments.length;
-      bigPictureDescription.textContent = picture.description;
-
     });
     fragment.appendChild(allPictures);
   });
